requirements.txt
python-telegram-bot
speechrecognition
pydub